Added property kb_force to class Enemy

+++ src/Entity/methods/startKnockback.cpp
@@ -2,7 +2,7 @@
 
 namespace Dungeon {
 
-void Entity::startKnockback(int angle, int force){
+void Entity::startKnockback(int angle, float force){
 	  this->kb_current_frame = 0; 
     this->kb_last_frame = force;
     this->kb_angle = angle;
